decouple Elasticsearch and Opensearch
use method override instead of func variables

fix mypy

default value

fix broken tests

update schema
