+++ haystack/nodes/retriever/sparse.py
@@ -8,7 +8,7 @@ from sklearn.feature_extraction.text import TfidfVectorizer
 
 from haystack.schema import Document
 from haystack.document_stores.base import BaseDocumentStore
-from haystack.document_stores.elasticsearch import KeywordDocumentStore
+from haystack.document_stores import KeywordDocumentStore
 from haystack.nodes.retriever import BaseRetriever
